requirements.txt
africastalking==1.2.7
aiohttp==3.9.0
aiohttp-retry==2.8.3
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
bcrypt==4.0.1
beanie==1.23.6
boto3==1.34.15
botocore==1.34.15
Brotli==1.1.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
cryptography==41.0.7
cssselect2==0.7.0
cups==0.0.6
Deprecated==1.2.14
dnspython==2.4.2
doc==0.1.0
ecdsa==0.18.0
email-validator==2.1.0.post1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
exponent-server-sdk==2.0.0
fastapi==0.104.1
firebase-admin==6.3.0
fonttools==4.45.1
frozenlist==1.4.0
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
hiredis==2.2.3
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joserfc==0.9.0
lazy-model==0.2.0
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
motor==3.3.2
msgpack==1.0.7
multidict==6.0.4
munch==4.0.0
numpy==1.26.2
openpyxl==3.1.2
orjson==3.9.10
packaging==23.2
panda==0.3.1
pandas==2.1.3
passlib==1.7.4
pdfkit==1.0.0
phonenumbers==8.13.26
Pillow==10.1.0
platformdirs==4.1.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.1
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.3
pydyf==0.8.0
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.0
pyotp==2.9.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyphen==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.4
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rich==13.7.0
rsa==4.9
s3transfer==0.10.0
schema==0.7.5
shellingham==1.5.4
shortuuid==1.0.11
six==1.16.0
smile-id-core==2.1.5
sniffio==1.3.0
starlette==0.27.0
stripe==7.9.0
tablib==3.6.0
text-unidecode==1.3
tinycss2==1.2.1
toml==0.10.2
twilio==8.10.2
typer==0.9.0
types-requests==2.31.0.20240403
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
weasyprint==60.1
webencodings==0.5.1
websockets==12.0
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.9.3
zeep==4.2.1
zopfli==0.2.3
ipython
holidays